Декомозировал компонент. Выделил в отдельный компонент каталог с жанрами и карточками фильмов
